+++ flinkboot-core/src/main/java/io/github/sekelenao/flinkboot/core/internal/execution/customizer/StateBackendCustomizer.java
@@ -28,7 +28,6 @@ public final class StateBackendCustomizer implements EnvironmentCustomizer {
     private void apply(StateBackendProperties stateConfig) {
         stateConfig.type().ifPresent(type -> this.applyType(type, stateConfig));
         stateConfig.checkpointStorage().ifPresent(this::applyCheckpointStorage);
-        stateConfig.storagePath().ifPresent(this::applyStoragePath);
         stateConfig.incremental().ifPresent(this::applyIncremental);
         stateConfig.latencyTracking().ifPresent(this::applyLatencyTracking);
     }
@@ -53,10 +52,6 @@ public final class StateBackendCustomizer implements EnvironmentCustomizer {
     private void applyCheckpointStorage(CheckpointStorageType checkpointStorage) {
         var storageName = checkpointStorage.toString().toLowerCase();
         toConfigure.set(CheckpointingOptions.CHECKPOINT_STORAGE, storageName);
-    }
-
-    private void applyStoragePath(String storagePath) {
-        toConfigure.set(CheckpointingOptions.CHECKPOINTS_DIRECTORY, storagePath);
     }
 
     private void applyIncremental(boolean incremental) {